price-writer/requirements.txt
kafka-python==2.0.2
zstandard==0.22.0
